Fix use of color channel keywords in calc() division denominators
The calc expression parser was enforcing a restriction that the
denominator must be a number, and depending on this restriction to
simplify divisions at parse time. This restriction was specified in
css-values-3 but relaxed in css-values-4.

Prior to crrev.com/c/371224775, when a color channel keyword was used
in a calc() expression, we would eagerly substitute in the numeric value
as soon as we encountered the keyword token. As a result, the parsed
expression would satisfy the restriction and the expression would be
accepted. As of that change, we now do the value substitution later in
order to preserve the originally specified form.

The fix is to relax the type restriction in code, which necessitates
implementing support for Invert nodes in calculation expression trees
as specified in css-values-4.

Bug: 371224775
Change-Id: Ia9b08030a08a64023b166abbd332e8c6db9043d0